cambio de tipo de dato de monto en Prestamo: double a BigDecimal

+++ src/main/java/com/banco/evaluacion/model/Prestamo.java
@@ -1,5 +1,7 @@
 package com.banco.evaluacion.model;
 
+import java.math.BigDecimal;
+
 //Para el prestamo no nos complicaremos será algo sencillo, de momento trabajaremos con soles, luego agregaremos otras monedas y haremos los calculos necesarios
-public record Prestamo(int id, int cliente_id, double monto, int plazoMeses, TipoPrestamo tipoPrestamo, EstadoPrestamo estado) {
+public record Prestamo(int id, int cliente_id, BigDecimal monto, int plazoMeses, TipoPrestamo tipoPrestamo, EstadoPrestamo estado) {
 }
